improves code maintanability
